requirements.txt
feedparser>=6.0
requests>=2.28
beautifulsoup4>=4.12
lxml>=4.9
pyyaml>=6.0
python-dateutil>=2.8
jinja2>=3.1

# Optional: LLM summarization
# openai>=1.0
# anthropic>=0.20
